device_manager: Default to null for current_song
saken-server-0.6a started to sync this field with the filename
field of the songs table. Meaning that the value here must either be
null or a valid song reference. Since I haven't really implemented
the option to update that side of the table, let's just return
null here for the time being.